requirements.txt
fastapi
uvicorn[standard]
requests

google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib      

openai>=1.3                    
twilio                   
python-dotenv             
gspread
oauth2client
